fix(contract): education 的三处 score 与 quality 的一处键序（字节比对报出）
第四轮捕获（2945 passed, 0 failed）与 education/quality 的"改前"基线逐字节
比对，78 个组合里报出 4 处真差异（另 2 处是并发报名测试里谁抢到名额不固定，
属噪声）：

1. **三处 `score` 被 int 截断**：`training_records.score` 与
   `training_assessments.score` 都是 FLOAT 列，我按"考核分数像整数"声明成
   int，`55.0` 被输出成 `55`、`88.0` 输出成 `88`。

   这是本轮第二次栽在同一个词上——"分数"在本仓库里三次都是 FLOAT 列
   （spd 量表得分、课程考核、实训考核）。已在治理文档里写成一条通则：
   遇到 score / rate / amount 这类"看起来像整数"的字段，一律先看列类型。

2. **一处键序被继承改掉**：`GET /api/quality/records/{id}/qc` 的 handler 出的是
   `{"record_id": ..., **result}`，record_id 在最前；我用继承建模
   （`RecordRescoreOut(RecordQcResultOut)`），父类字段先排、子类新增字段在后，
   输出变成了 `{score, ..., record_id}`。

   改法是逐字段重列、顺序照 handler。继承只在"多出来的字段本来就在末尾"时
   才安全——这条作为「陷阱五」补进治理文档。

两处都已实测复核：exam 回 `55.0`（float），复评键序回到 record_id 在最前。
ruff / mypy 全过；education/quality 涉及的四个测试文件整文件复跑 42 passed。

+++ server/app/routers/education.py
@@ -54,7 +54,9 @@ class CourseOut(CourseCreate):
 # 响应契约集中放在所有端点之前（`response_model=` 是装饰器参数，导入时求值）
 class ExamResultOut(BaseModel):
     course_id: int
-    score: int
+    # `training_records.score` 是 FLOAT 列——声明成 int 会把 55.0 输出成 55。
+    # 字节比对实测到的（与 spd 的量表得分同一形状：分数听起来像整数，列不是）。
+    score: float
     passed: bool
 
 
@@ -69,7 +71,7 @@ class CourseStatsOut(BaseModel):
 class MyRecordOut(BaseModel):
     course_id: int
     title: str
-    score: int
+    score: float
     passed: bool
 
 
@@ -179,7 +181,8 @@ class AssessmentOut(BaseModel):
     id: int
     plan_id: int
     user_id: int
-    score: int
+    # `training_assessments.score` 同样是 FLOAT
+    score: float
     passed: bool
     assessor: str
 
@@ -187,7 +190,7 @@ class AssessmentOut(BaseModel):
 class AssessmentRowOut(BaseModel):
     id: int
     user_id: int
-    score: int
+    score: float
     passed: bool
     comment: str
     assessor: str
